Move !stats, !time, !admins, !goat and !fgoat to another module
